feature-react-router: implemented absolute imports and integrated react router
