+++ src/components/Promotion/Card/Card.js
@@ -2,7 +2,11 @@ import './Card.css';
 
 const PromotionCard = ({ promotion }) => (
   <div className="promotion-card">
-    <img src={promotion.imageUrl} className="promotion-card__image" />
+    <img
+      src={promotion.imageUrl}
+      alt={promotion.title}
+      className="promotion-card__image"
+    />
     <div className="promotion-card__info">
       <h1 className="promotion-card__title">{promotion.title}</h1>
       <span className="promotion-card__price">R$ {promotion.price}</span>
@@ -19,6 +23,7 @@ const PromotionCard = ({ promotion }) => (
         <a
           href={promotion.url}
           target="_blank"
+          rel="noreferrer"
           className="promotion-card__link"
         >
           IR PARA O SITE
