+++ src/pages/Promotion/Form/Form.js
@@ -1,0 +1,13 @@
+import { useParams } from 'react-router-dom';
+
+const PagesPromotionForm = () => {
+  const { id } = useParams();
+  return (
+    <div>
+      Form
+      {id && <div>id: {id}</div>}
+    </div>
+  );
+};
+
+export default PagesPromotionForm;
